example: break-ligatures: Refrain from concatenating discretionary pre and post lists.
In a discretionary, there is not necessarily a strong relation between
pre, post and replace lists.  Contructing a replace list as the
(augmented) concatenation of pre and post lists can therefore give
wrong results.  A ligature in a replace list is now broken by
manipulating the original replace list in all cases.